Move some functions from SpanMatcher to a companion object.
